voltage.h: get_voltage returns unsigned long

+++ arch/arm/mach-omap2/voltage.h
@@ -92,7 +92,7 @@ struct voltagedomain {
 	 * FIXME should this be a long to return error codes?
 	 * e.g. if VC cannot support get_voltage, return -EPERM
 	 */
-	u32 (*get_voltage) (struct voltagedomain *voltdm);
+	unsigned long (*get_voltage) (struct voltagedomain *voltdm);
 
 	u32 nominal_volt;
 	struct omap_volt_data *volt_data;
